CachedSchemaRegistryClient -> throws on cache full. other refactoring
