+++ src/Confluent.SchemaRegistry/Rest/DataContracts/ErrorMessage.cs
@@ -17,7 +17,7 @@
 using System.Runtime.Serialization;
 
 
-namespace Confluent.SchemaRegistry.Rest.Entities
+namespace Confluent.SchemaRegistry
 {
     /// <summary>
     ///     Generic JSON error message.
