+++ src/Confluent.SchemaRegistry/Rest/DataContracts/Schema.cs
