+++ test/Confluent.SchemaRegistry.IntegrationTests/Tests/FillTheCache.cs
@@ -46,35 +46,27 @@ namespace Confluent.SchemaRegistry.IntegrationTests
 
             var subjects = new List<string>();
             var ids = new List<int>();
-            for (int i=0; i<registerCount; ++i)
+
+            Assert.Throws<OutOfMemoryException>(() =>
             {
-                var topicName = Guid.NewGuid().ToString();
-                var subject = sr.ConstructValueSubjectName(topicName);
-                subjects.Add(subject);
+                try
+                {
+                    for (int i = 0; i < registerCount; ++i)
+                    {
+                        var topicName = Guid.NewGuid().ToString();
+                        var subject = sr.ConstructValueSubjectName(topicName);
+                        subjects.Add(subject);
 
-                var id = sr.RegisterAsync(subject, testSchema).Result;
-                ids.Add(id);
-            }
+                        var id = sr.RegisterAsync(subject, testSchema).Result;
+                        ids.Add(id);
+                    }
+                }
+                catch (AggregateException e)
+                {
+                    throw e.InnerException;
+                }
+            });
 
-            for (int i=0; i<registerCount; ++i)
-            {
-                var s = sr.GetSchemaAsync(ids[i]).Result;
-                Assert.Equal(s, testSchema);
-            }
-
-            for (int i=0; i<registerCount; ++i)
-            {
-                var latest = sr.GetLatestSchemaAsync(subjects[i]).Result;
-                Assert.True(latest.Id > 0);
-                Assert.Equal(latest.Version, 1);
-                Assert.Equal(latest.SchemaString, testSchema);
-                Assert.Equal(latest.Subject, subjects[i]);
-
-                var schema = sr.GetSchemaAsync(subjects[i], latest.Version).Result;
-                Assert.NotNull(schema);
-            }
-
-            // Note: Inspect in debugger for detailed internal cache behavior investigation.
         }
     }
 }
